Initial EL7 packaging
libvarlink is already shipped in Fedora, rationale in repackaging it for EL7
lies in the fact, it uses meson as a build system. Meson is a python3-only
component, so we had to ressurect the former autotools-based build-system
and fix it to build current versions.
So no new reviews is required.